the start of a new project
hi! i was working in a new project, i was busy with the university so i can have much time but i will end this project soon!

my upgrade was all this, the creation a server, have all routes, do validators, create a connection to database and more!

i will try upgrade my code if it need and keep goin in my project :)